+++ Changelog.md
@@ -12,6 +12,7 @@
 - Add pod-reaper enabled label to PingCentral
 - Upgrade fluent-bit to version 3.2.10
 - Logstash: Revise Bootstrap Process to Directly Check Status of OpenSearch
+- Add p1as-Observability to PCB Grafana
 
 _Changes:_
 
@@ -19,6 +20,7 @@ _Changes:_
 - [X] PDO-8754 OSD: Enable Save Query and CSV Export of Logs
 - [X] PDO-8834 Update bootstrap pods to use new bootstrap IRSA role
 - [X] PDO-9021 Enable p1as-cluster-tools microservice in PCB
+- [X] PDO-9231 Add p1as-observability to PCB: Grafana
 - [X] PDO-9312 Get pingone URL from env variable
 - [X] PDO-9376 FluentBit: Adjust Buffer and Limit Configuration
 - [X] PDO-9414 Set initialDelaySeconds and failureThreshold on readiness probe for PingAccess and